Update README.rst
As I was following the tutorial and running the examples, I was confused as to how I should be able to view the metrics in my local MLflow Tracking URI. I could see that the UI was running in http://localhost/5000 and the logistic regression model was also executed, but was confused as to why I could not view the metrics. If we mention how to view the runs logged as part of the README, it can help other users who may also have face this confusion.